make clsim work with ancient versions of IceTray
git-svn-id: http://code.icecube.wisc.edu/svn/sandbox/ckopper/clsim/trunk@86828 16731396-06f5-0310-8873-f7f720988828

+++ private/clsim/I3CLSimModule.cxx
@@ -806,7 +806,7 @@ void I3CLSimModule::AddPhotonsToFrames(const I3CLSimPhotonSeries &photons,
             
             if (photonHistory.size() > photon.GetNumScatters())
                 log_fatal("Logic error: photonHistory.size() [==%zu] > photon.GetNumScatters() [==%zu]",
-                          photonHistory.size(), photon.GetNumScatters());
+                          photonHistory.size(), static_cast<std::size_t>(photon.GetNumScatters()));
             
             for (std::size_t j=0;j<photonHistory.size();++j)
             {
